rename stub file.
